+++ src/TextShow/Data/Fixed.hs
@@ -25,7 +25,6 @@ import TextShow.Classes (TextShow(..))
 #if MIN_VERSION_base(4,7,0)
 import Data.Fixed (Fixed(..))
 import Data.Int (Int64)
-import Data.Monoid.Compat ((<>))
 import Data.Semigroup (mtimesDefault)
 import Data.Text.Lazy.Builder (singleton)