+++ src/TextShow/Generic.hs
@@ -89,7 +89,6 @@ module TextShow.Generic (
 
 import           Data.Data (Data, Typeable)
 import           Data.Functor.Contravariant (Contravariant(..))
-import           Data.Monoid.Compat ((<>))
 import qualified Data.Text    as TS (Text, pack, singleton)
 import qualified Data.Text.IO as TS (putStrLn, hPutStrLn)
 import qualified Data.Text.Lazy    as TL (Text, pack, singleton)