+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+### next [????.??.??]
+* Use `base-compat-batteries`.
+
 ### 3.7.2 [2018.03.08]
 * Depend on `quickcheck-instances-0.3.18` or later.